Setting up long feature occasion setting experiment
Shouldn't have to change Nept this often for these types of
MedPC experiments. Need to re-evaluate how I do the analysis.
The objects make things easy when they're working, but difficult
to extract out other information.

+++ nept/utils.py
@@ -75,7 +75,7 @@ def consecutive(array, stepsize):
     Parameters
     ----------
     array : np.array
-    stepsize : int
+    stepsize : float
 
     Returns
     -------
